Add host SSH fingerprints to documentation
When connecting to an SSH server for the first time, some more security-conscious users may wish to ensure that they are in fact connecting to the correct remote host and not a MITM (man in the middle) machine. This is especially important for the first connection since the fingerprint(s) learned will likely be used to verify that future connections are genuine. This commit adds user-facing host fingerprints to the documentation to facilitate cross-checking by those who so wish.